ai110-week8-project/requirements.txt
langchain>=0.3.0
langchain-google-genai>=2.0.0
langchain-community>=0.3.0
faiss-cpu>=1.7.0
streamlit>=1.38.0
pypdf>=4.0.0
google-generativeai>=0.8.0
python-dotenv>=1.0.0
pytest
